+++ tests/test_views.py
@@ -1,16 +1,25 @@
+from unittest.mock import patch
+
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.sites.models import Site
+from django.test import override_settings
 
-from cms.models import Page, User
+from cms.models import Page, PageContent, User
 from cms.test_utils.testcases import CMSTestCase
+from cms.utils import get_current_site
 from cms.utils.urlutils import admin_reverse
 
+from djangocms_versioning.constants import PUBLISHED
+
 from djangocms_navigation.constants import SELECT2_CONTENT_OBJECT_URL_NAME
+from djangocms_navigation.models import MenuContent
 from djangocms_navigation.test_utils.factories import (
     MenuContentFactory,
     PageContentFactory,
+    PageContentWithVersionFactory,
 )
 from djangocms_navigation.test_utils.polls.models import Poll, PollContent
+from djangocms_navigation.views import ContentObjectSelect2View
 
 
 class PreviewViewPermissionTestCases(CMSTestCase):
@@ -127,6 +136,7 @@ class ContentObjectAutoFillTestCases(CMSTestCase):
         self.assertEqual(response.status_code, 200)
         expected_list = [{"text": "test", "id": 1}, {"text": "test2", "id": 2}]
         self.assertEqual(response.json()["results"], expected_list)
+        self.assertEqual(len(response.json()["results"]), 2)
 
     def test_select2_view_search_text_page(self):
         """ Both pages should appear in results for test query"""
@@ -214,3 +224,242 @@ class ContentObjectAutoFillTestCases(CMSTestCase):
         self.assertEqual(response.status_code, 200)
         expected_json = {"results": [{"text": "example", "id": 1}]}
         self.assertEqual(response.json(), expected_json)
+
+    @override_settings(DJANGOCMS_NAVIGATION_VERSIONING_ENABLED=True)
+    def test_with_multiple_versions_distinct_results_returned(self):
+        """
+        Check that when there are multiple Pages, and each have multiple versions of PageContent, that the returned
+        Page objects are distinct and do not contain duplicate titles/ids
+        """
+        page_contenttype_id = ContentType.objects.get_for_model(Page).id
+        first_page = PageContentWithVersionFactory(
+            title="test", menu_title="test", page_title="test", language="en", version__state=PUBLISHED
+        )
+        # create a draft version of the page
+        first_page_new_version = first_page.versions.get().copy(self.superuser)
+        first_page_new_version.save()
+        second_page = PageContentWithVersionFactory(
+            title="test2", menu_title="test2", page_title="test2", language="en"
+        )
+        # create a draft version and publish it
+        second_page_new_version = second_page.versions.get().copy(self.superuser)
+        second_page_new_version.save()
+        second_page_new_version.publish(self.superuser)
+
+        with self.login_user_context(self.superuser):
+            response = self.client.get(
+                self.select2_endpoint,
+                data={"content_type_id": page_contenttype_id, "query": "test"},
+            )
+            results = response.json()["results"]
+
+        self.assertEqual(Page._base_manager.count(), 2)
+        self.assertEqual(PageContent._base_manager.count(), 4)
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(len(results), 2)
+        expected = [
+            {"text": "test", "id": first_page.page.pk},
+            {"text": "test2", "id": second_page.page.pk}
+        ]
+        self.assertEqual(results, expected)
+
+    def test_with_pages_in_multiple_languages(self):
+        """
+        Check that when page content exists in multiple languages, only pages for the current language are returned
+        """
+        page_contenttype_id = ContentType.objects.get_for_model(Page).id
+        french = PageContentFactory(
+            title="test", menu_title="test", page_title="test", language="fr",
+        )
+        PageContentFactory.create_batch(10, title="test", menu_title="test", page_title="test", language="en")
+
+        with self.login_user_context(self.superuser):
+            url = self.select2_endpoint.replace("/en/", "/fr/")
+            response = self.client.get(
+                url,
+                data={"content_type_id": page_contenttype_id, "query": "test"},
+            )
+            results = response.json()["results"]
+
+        self.assertEqual(PageContent._base_manager.count(), 11)
+        self.assertEqual(len(results), 1)
+        expected = [
+            {"text": "test", "id": french.page.pk},
+        ]
+        self.assertEqual(results, expected)
+
+    def test_with_pages_for_multiple_sites(self):
+        """
+        Check that with pages for multiple sites, only pages for the requested site are returned
+        """
+        page_contenttype_id = ContentType.objects.get_for_model(Page).id
+        site1 = Site.objects.create(name="site1.com", domain="site1.com")
+        site2 = Site.objects.create(name="site2.com", domain="site2.com")
+        PageContentFactory.create_batch(10, title="test", page__node__site=site1, language="en")
+        expected = PageContentFactory(title="test", menu_title="site2 page", page__node__site=site2, language="en")
+
+        with self.login_user_context(self.superuser):
+            response = self.client.get(
+                self.select2_endpoint,
+                data={
+                    "content_type_id": page_contenttype_id,
+                    "site": site2.id
+                },
+            )
+            results = response.json()["results"]
+
+        self.assertEqual(PageContent._base_manager.count(), 11)
+        self.assertEqual(len(results), 1)
+        self.assertEqual(results, [{"id": expected.page.id, "text": "site2 page"}])
+
+
+class ContentObjectSelect2ViewGetDataTestCase(CMSTestCase):
+
+    def setUp(self):
+        """
+        Setup a view object with a request without a search query that can be used in unit tests.
+        The request object can be modified as required for the unit test.
+        """
+        self.page_contenttype_id = ContentType.objects.get_for_model(Page).id
+        self.request = self.get_request(language="en")
+        self.view = ContentObjectSelect2View(request=self.request, menu_content_model=MenuContent)
+
+    @patch("django.db.models.QuerySet.distinct")
+    def test_distinct_not_called_without_search_query_for_poll(self, mock_distinct):
+        """
+        Mock distinct to assert that it is not called if there is no search query in the request for the poll content
+        type
+        """
+        poll_content_contenttype_id = ContentType.objects.get_for_model(PollContent).id
+        self.request.GET = {"content_type_id": poll_content_contenttype_id, "query": None}
+
+        self.view.get_data()
+
+        mock_distinct.assert_not_called()
+
+    @patch("django.db.models.QuerySet.distinct")
+    def test_distinct_called_with_search_query_for_page(self, mock_distinct):
+        """
+        Mock distinct to assert that it is called for the page content type even without a search query, because the
+        queryset is filtered by language and site
+        """
+        self.request.GET = {"content_type_id": self.page_contenttype_id, "query": None}
+
+        self.view.get_data()
+
+        mock_distinct.assert_called_once()
+
+    @patch("django.db.models.QuerySet.distinct")
+    def test_distinct_called_with_search_query(self, mock_distinct):
+        """
+        Mock distinct to assert that it is called if there is a search query in the request
+        """
+        self.request.GET = {"content_type_id": self.page_contenttype_id, "query": "test query"}
+
+        self.view.get_data()
+
+        mock_distinct.assert_called_once()
+
+    def test_results_unfiltered_without_search_fields(self):
+        """
+        Check that if no search fields have been declared to filter against, the queryset is returned unfiltered
+        """
+        poll_content_contenttype_id = ContentType.objects.get_for_model(PollContent).id
+        poll = Poll.objects.create(name="Test poll")
+        poll1 = PollContent.objects.create(poll=poll, language="en", text="poll1")
+        poll2 = PollContent.objects.create(poll=poll, language="en", text="poll2")
+        self.request.GET = {"content_type_id": poll_content_contenttype_id, "query": "poll1"}
+
+        with patch("djangocms_navigation.views.supported_models") as mock:
+            mock.return_value = {PollContent: []}
+            results = self.view.get_data()
+
+        self.assertEqual(results.count(), 2)
+        self.assertIn(poll1, results)
+        self.assertIn(poll2, results)
+
+    def test_results_filtered_by_search_fields(self):
+        """
+        Check that if the search fields has been declared to filter against, the queryset is returned filtered
+        """
+        poll_content_contenttype_id = ContentType.objects.get_for_model(PollContent).id
+        poll = Poll.objects.create(name="Test poll")
+        poll1 = PollContent.objects.create(poll=poll, language="en", text="poll1")
+        poll2 = PollContent.objects.create(poll=poll, language="en", text="poll2")
+        self.request.GET = {"content_type_id": poll_content_contenttype_id, "query": "poll1"}
+
+        with patch("djangocms_navigation.views.supported_models") as mock:
+            mock.return_value = {PollContent: ["text"]}
+            results = self.view.get_data()
+
+        self.assertEqual(results.count(), 1)
+        self.assertIn(poll1, results)
+        self.assertNotIn(poll2, results)
+
+    def test_page_queryset_is_filtered_by_content_title(self):
+        """
+        Check that the returned Page QuerySet is filtered by the PageContent title
+        """
+        expected = PageContentFactory(title="Example PageContent", language="en")
+        PageContentFactory.create_batch(10)
+        self.request.GET = {"content_type_id": self.page_contenttype_id, "query": expected.title}
+
+        results = self.view.get_data()
+
+        self.assertEqual(Page._base_manager.count(), 11)
+        self.assertEqual(results.count(), 1)
+        self.assertIn(expected.page, results)
+
+    def test_page_queryset_filters_pages_by_current_language(self):
+        """
+        Check that only the page for the language of the request is returned
+        """
+        self.en = PageContentFactory(title="english test", language="en")
+        self.fr = PageContentFactory(title="french test", language="fr")
+        self.de = PageContentFactory(title="german test", language="de")
+        self.it = PageContentFactory(title="italian test", language="it")
+
+        for language in ["en", "fr", "de", "it"]:
+            with self.subTest(msg=language):
+                request = self.get_request(language=language)
+                request.GET = {"content_type_id": self.page_contenttype_id, "query": "test"}
+                self.view.request = request
+                results = self.view.get_data()
+
+                expected = getattr(self, language)
+                self.assertEqual(Page._base_manager.count(), 4)
+                self.assertEqual(results.count(), 1)
+                self.assertIn(expected.page, results)
+
+    @patch("djangocms_navigation.views.get_current_site")
+    def test_page_queryset_filtered_by_site(self, mock_get_current_site):
+        """
+        Check that when pages belong to different sites, the returned queryset only includes pages for the current site
+        """
+        site1 = Site.objects.create(domain="site1.com", name="site1")
+        site2 = Site.objects.create(domain="site2.com", name="site2")
+        PageContentFactory.create_batch(10, language="en", page__node__site=site1)
+        PageContentFactory.create_batch(10, language="en", page__node__site=site2)
+
+        for site in [site1, site2]:
+            with self.subTest(msg=site.name):
+                self.request.GET = {"content_type_id": self.page_contenttype_id}
+                mock_get_current_site.return_value = site
+                results = self.view.get_data()
+
+                expected_domain = results.values_list("node__site__domain", flat=True).distinct()
+                self.assertEqual(Page._base_manager.count(), 20)
+                self.assertEqual(results.count(), 10)
+                self.assertEqual(expected_domain.first(), site.domain)
+
+    def test_when_no_site_param_get_current_site_called(self):
+        """
+        Check that when no site param is included in the request, that the get_current_site method is called
+        """
+        self.request.GET = {"content_type_id": self.page_contenttype_id, "site": None}
+
+        # use wraps to allow us to assert the method is called while retaining the original behaviour of the method
+        with patch("djangocms_navigation.views.get_current_site", wraps=get_current_site) as mock_get_current_site:
+            self.view.get_data()
+
+        mock_get_current_site.assert_called_once()