+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 
 Unreleased
 ==========
+* fix: Remove duplicate objects when adding MenuItem nodes
+* feat: When adding a MenuItem the Page objects are filtered for the current site and language
 
 1.5.0 (2022-08-22)
 ==================